specify dtype for linreg

+++ xftsim/stats.py
@@ -550,7 +550,7 @@ def _mv_vec_linear_regression_with_intercept_nb(X: NDArray,
                                                 Y: NDArray,
                                                 std_X: bool = True,
                                                 std_Y: bool = True,
-                                                ) -> NDArray:
+                                                output_dtype=np.float64) -> NDArray:
     """
     Numba implementation of simple linear regression vectorized over predictors
     and outcomes. 
@@ -576,7 +576,7 @@ def _mv_vec_linear_regression_with_intercept_nb(X: NDArray,
     #     for j in np.arange(X.shape[1]):
     #         output[j,:, k] = _linear_regression_with_intercept_nb(X[:,j],y)
     # return output
-    output = np.empty((X.shape[1], 2, Y.shape[1]), dtype=X.dtype)
+    output = np.empty((X.shape[1], 2, Y.shape[1]), dtype=output_dtype)
     for k in nb.prange(Y.shape[1]):
         y = Y[:,k].ravel()
         y = y.reshape((y.shape[0], 1))
